ai_engine/requirements.txt
torch
torchvision
transformers
pillow
numpy
pandas
scikit-learn
tqdm
